Moved FarmMapsAuthorizationMessageHandler class
